Move copy service to gaphor.ui
