+++ gaphor/ui/copyservice.py
