+++ pyproject.toml
@@ -119,7 +119,7 @@ gaphorconvert = "gaphor.plugins.diagramexport.gaphorconvert:main"
 "main_window" = "gaphor.ui.mainwindow:MainWindow"
 "export_menu" = "gaphor.ui.menufragment:MenuFragment"
 "tools_menu" = "gaphor.ui.menufragment:MenuFragment"
-"copy" = "gaphor.services.copyservice:CopyService"
+"copy" = "gaphor.ui.copyservice:CopyService"
 "sanitizer" = "gaphor.UML.sanitizerservice:SanitizerService"
 "model_browser" = "gaphor.ui.modelbrowser:ModelBrowser"
 "toolbox" = "gaphor.ui.toolbox:Toolbox"
